+++ src/commands/bub.js
@@ -162,11 +162,9 @@ const bub = async (intCmds, file, ...args) => {
     if (err.code === "EPERM") {
       Verbose.permError();
       Errors.noPermissions("read the file", file);
-      return;
     } else if (err.code === "EBUSY") {
       Verbose.inUseError();
       Errors.inUse("file", file);
-      return;
     } else {
       Verbose.fatalError();
       _fatalError(err);