+++ src/commands/cd.js
@@ -93,7 +93,15 @@ const cd = (dir, ...args) => {
     if (err.code === "EPERM") {
       Verbose.permError();
       Errors.noPermissions("change into", dir);
-      return;
+    } else if (err.code === "ENOENT") {
+      // For some reason, there are rare cases where the checks think the directory exists,
+      // but when trying to change into it, it throws an error.
+      // This usually happens when using the BubbleOS executable, where a folder called
+      // "C:\snapshot" is visible on Windows (through 'ls' in BubbleOS), but when trying to
+      // change into it, it throws an error.
+
+      Verbose.custom("Directory does not exist.");
+      Errors.doesNotExist("directory", dir);
     } else {
       Verbose.fatalError();
       _fatalError(err);