+++ src/commands/hash.js
@@ -124,11 +124,9 @@ const hash = (file, ...args) => {
     if (err.code === "EPERM") {
       Verbose.permError();
       Errors.noPermissions("read the file", file);
-      return;
     } else if (err.code === "EBUSY") {
       Verbose.inUseError();
       Errors.inUse("file", file);
-      return;
     } else {
       Verbose.fatalError();
       _fatalError(err);