+++ src/commands/ls.js
@@ -163,7 +163,15 @@ const ls = (dir = `"${process.cwd()}"`, ...args) => {
     if (err.code === "EPERM") {
       Verbose.permError();
       Errors.noPermissions("read", dir);
-      return;
+    } else if (err.code === "ENOENT") {
+      // For some reason, there are rare cases where the checks think the directory exists,
+      // but when trying to list the contents, it throws an error.
+      // This usually happens when using the BubbleOS executable, where a folder called
+      // "C:\snapshot" is visible on Windows (through 'ls' in BubbleOS), but when trying to
+      // list it, it throws an error.
+
+      Verbose.custom("Directory does not exist.");
+      Errors.doesNotExist("directory", dir);
     } else {
       Verbose.fatalError();
       _fatalError(err);