+++ src/commands/print.js
@@ -1,3 +1,5 @@
+const _fatalError = require("../functions/fatalError");
+
 const Checks = require("../classes/Checks");
 const Verbose = require("../classes/Verbose");
 
@@ -7,16 +9,21 @@ const Verbose = require("../classes/Verbose");
  * @param {...string} text The text to output.
  */
 const print = (...text) => {
-  // If no text passed, print nothing at all
-  if (new Checks(text).paramUndefined()) {
-    Verbose.chkEmpty();
-    console.log();
-    return;
-  }
+  try {
+    // If no text passed, print nothing at all
+    if (new Checks(text).paramUndefined()) {
+      Verbose.chkEmpty();
+      console.log();
+      return;
+    }
 
-  // Print the text, and end it with a newline
-  Verbose.custom("Printing text...");
-  console.log(`${text.join(" ")}\n`);
+    // Print the text, and end it with a newline
+    Verbose.custom("Printing text...");
+    console.log(`${text.join(" ")}\n`);
+  } catch (err) {
+    Verbose.fatalError();
+    _fatalError(err);
+  }
 };
 
 module.exports = print;