+++ src/commands/readfile.js
@@ -122,11 +122,9 @@ const readfile = (file, ...args) => {
     if (err.code === "EPERM") {
       Verbose.permError();
       Errors.noPermissions("read the file", file);
-      return;
     } else if (err.code === "EBUSY") {
       Verbose.inUseError();
       Errors.inUse("file", file);
-      return;
     } else {
       Verbose.fatalError();
       _fatalError(err);