+++ CHANGELOG.md
@@ -15,6 +15,8 @@ This is the official BubbleOS changelog! All features will be recorded in _'grou
 - Fixed an issue where the `ping` command would hang BubbleOS if the status code was 200 or the server timed out. BubbleOS would also terminate if the status code was not 200.
 - Fixed an issue where the `ls` command would crash BubbleOS if it did not have permissions to read the directory.
 - Fixed an issue where the battery percentage would sometimes show as `NaN%` if the device does not have a battery (mainly affected Windows devices).
+- Fixed an issue where the `cd` and `ls` would rarely crash BubbleOS if the directory did not exist.
+- Fixed an issue where the `ifnet` command did not have any verbose messages.
 
 ### Removed Features